+++ types/cupcake.go
@@ -4,6 +4,6 @@ type Cupcake struct {
 	Ingredient IngredientAdder
 }
 
-func (c *Cupcake) AddIngredient() (string, float32, error) {
-	return "Cupcake with", 1, nil
+func (c *Cupcake) AddIngredient() (string, float64, error) {
+	return "Cupcake with", 1.0, nil
 }
