+++ types/ingredient-adder.go
@@ -1,5 +1,5 @@
 package types
 
 type IngredientAdder interface {
-	AddIngredient() (string, float32, error)
+	AddIngredient() (string, float64, error)
 }
